+++ src/sgr/color/rgb.ts
@@ -14,7 +14,7 @@ export function makeRGB(ctx: Context): Channel["rgb"] {
 			channel === "fg" ?
 				(r, g, b) => x16ToX8(rgbToX16(...clip(r, g, b))) :
 				(r, g, b) => build(open + rgbToX8(...clip(r, g, b))) :
-		() => build(0)
+		() => build([])
 	);
 }
 
@@ -43,6 +43,6 @@ export function makeX256(ctx: Context): (...args: [number] | [number, number, nu
 						...args.length === 3 ? clip(...args) : x256ToRgb(clip(args[0])),
 					),
 				) :
-		() => build(0)
+		() => build([])
 	);
 }